% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpParallel.r
\name{endParallel}
\alias{endParallel}
\title{Function that completes the necessary steps to stop the parallel cluster started with setUpParallel()}
\usage{
endParallel(cl)
}
\arguments{
\item{c1}{A "cluster" object that was returned by setUpParallel().}
}
\value{
.
}
\description{
This function completes the necessary steps to stop the parallel cluster started with setUpParallel()
}
\examples{
endParallel (c1)
}
\keyword{foreach}
\keyword{parallel}
\keyword{stop}
