% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOptBounds.r
\name{getOptBounds}
\alias{getOptBounds}
\title{Function to resize the upper and lower boundary range of a parameter.}
\usage{
getOptBounds(arrayIn, originalUpperBound, originalLowerBound,
  minimumIntervalValue, numIntervals = 25)
}
\arguments{
\item{arrayIn}{An array of the the parameter's values from the best fit runs (presumably).}

\item{originalUpperBound}{An number that specifies the original, unaltered upper bound of the parameter that was input at the start of smartGridSearch.}

\item{originalLowerBound}{An number that specifies the original, unaltered lower bound of the parameter that was input at the start of smartGridSearch.}

\item{minimumIntervalValue}{An number that specifies the original, unaltered minimum interval for the parameter that was input at the start of smartGridSearch.}

\item{numIntervals}{An integer that specifies the number of intervals that the range between the upper and lower bounds of each parameter space will be segmented into.   DEFAULT = 25.}
}
\value{
An array of three values: "lower" = the new lower bound; "upper" = the new upper bound; "int" = the new interval value.
}
\description{
This function resize the upper and lower boundary range of a parameter based on the best fit runs of the simulation.  The resizing is done so that it iteratively narrows down to the best parameter values.
}
\examples{
getOptBounds(arrayIn = c(1,5,3,5,3), originalUpperBound = 10, originalLowerBound = 0, minimumIntervalValue = 0.1, numIntervals=25)
}
\keyword{bounds}
\keyword{optimal}
