% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOrderedParamList.r
\name{getOrderedParamList}
\alias{getOrderedParamList}
\title{Function to identify and return the top N runs of the simulation (sorted)}
\usage{
getOrderedParamList(paramList, listItemToMinimize = "r_out",
  optParamListN = 10, fn, otherParamList = NULL, optBoundLoops = 10)
}
\arguments{
\item{paramList}{A list of lists, with each sublist containing the parameter values for a run of the simulation.  The top level list contains all the runs of the simulation.}

\item{listItemToMinimize}{An string specifying the name of the list item that specifies the fit statistic being minimized.  This list item tells the program how well the simulation parameters fit the data. DEFAULT = "r_out" (which is what smartGridSearch uses)}

\item{optParamListN}{An integer that specifies the number of best fit runs that will be used to extract the new bounds for next set of grid search loops.  So, if optParamListN == 10, then the parameter values from the 10 best runs will be used to refine the upper and lower bounds of for each parameter. DEFAULT = 10.}

\item{fn}{A function that runs the model you would like to optimize.  It should return the fit statistic that is to be minimized.}

\item{otherParamList}{A list containing the other arguments/parameters to be passed to "fn."  The name of each element in the list must be the arguments/parameters name in "fn."}

\item{optBoundLoops}{An integer that specifies the number of times to re-simulate the best fit run in order to get a mean and SD for the minimized fit statistic. If the mean fit statistic is no longer the best fit, the program will move to the next best fit run, and so on until the it identifies the best mean fit run. DEFAULT = 10.}
}
\value{
A list of lists containing a sorted list of the 10 best runs with the sublists containing the parameter values of each run.
}
\description{
This function identifies the best N (optParamListN) runs of the simulation and returns them sorted by best run to worst run.
}
\examples{
getOrderedParamList(parameterListOfLists, "r_out", optParamListN = 10, fn, otherParamList, optBoundLoops = 10)
}
\keyword{fit}
\keyword{list}
\keyword{optimize}
\keyword{ordered}
\keyword{parameters}
\keyword{statistic}
