% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getRandomNumber.r
\name{getRandomNumber}
\alias{getRandomNumber}
\title{Function that returns "n" random numbers between an upper and lower bound.}
\usage{
getRandomNumber(n, min, max, interval)
}
\arguments{
\item{n}{An integer specifying the number of values to return.}

\item{min}{A number specifying the lower bound of the range to draw from.}

\item{max}{A number specifying the upper bound of the range to draw from.}

\item{interval}{A number specifying the size of the steps to take between the upper and lower bounds.}
}
\value{
n values between min and max, with an interval.
}
\description{
This Function that returns "n" random numbers between an upper and lower bound, with a specific interval.
}
\examples{
getRandomNumber (1, 0, 10, 1)
}
\keyword{max}
\keyword{min}
\keyword{number}
\keyword{random}
