% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setUpParallel.r
\name{setUpParallel}
\alias{setUpParallel}
\title{Function that initiates the necessary steps so that one can use "foreach" to run on multiple cores}
\usage{
setUpParallel()
}
\arguments{
\item{No}{parameters.}
}
\value{
on object of class "cluster"
}
\description{
This function initiates the necessary steps so that one can use "foreach" to run on multiple cores
}
\examples{
setUpParallel ()
}
\keyword{foreach}
\keyword{parallel}
\keyword{set}
\keyword{up}
