% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartGridSearch.r
\name{smartGridSearch}
\alias{smartGridSearch}
\title{Grid Search Optimization Function}
\usage{
smartGridSearch(
  fn,
  parsUpper,
  parsLower,
  parsMinInt,
  otherParamList = NULL,
  numLoops = 200,
  numIntervals = 25,
  optParamListN = 10,
  optBoundLoops = 10,
  roundPrecision = 3,
  progress = T,
  multicore = F,
  multicorePackages = NULL
)
}
\arguments{
\item{fn}{A function that runs the model you would like to optimize.  It should return the fit statistic that is to be minimized.}

\item{parsUpper}{A list containing, for each parameter, a vector of values for the upper bounds of that parameter's elements.  The name of the vector must correspond to the parameter name in the "fn."}

\item{parsLower}{A list containing, for each parameter, a vector of values for the lower bounds of that parameter's elements.  The name of the vector must correspond to the parameter name in the "fn."}

\item{parsMinInt}{A list containing, for each parameter, a vector of values for the minimum interval sizes to be used when randomly choosing a value between the upper and lower bounds of that parameter's elements.  The name of the vector must correspond to the parameter name in the "fn."}

\item{otherParamList}{A list containing the other arguments/parameters to be passed to "fn."  The name of each element in the list must be the arguments/parameters name in "fn."}

\item{numLoops}{An integer that specifies the number of loops with different parameter values to be searched in each run of the smartGridSearch before the best fits for that run are identified. DEFAULT = 200.}

\item{numIntervals}{An integer that specifies the number of intervals that the range between the upper and lower bounds of each parameter space will be segmented into.   DEFAULT = 25.}

\item{optParamListN}{An integer that specifies the number of best fit runs that will be used to extract the new bounds for next set of grid search loops.  So, if optParamListN == 10, then the parameter values from the 10 best runs will be used to refine the upper and lower bounds of for each parameter. DEFAULT = 10.}

\item{optBoundLoops}{An integer that specifies the number of times to re-simulate the best fit run in order to get a mean and SD for the minimized fit statistic. If the mean fit statistic is no longer the best fit, the program will move to the next best fit run, and so on until the it identifies the best mean fit run. DEFAULT = 10.}

\item{roundPrecision}{An integer that specifies the number of digits that the fit statistic should be rounded to when comparing determining if the optimization is complete. DEFAULT = 3.}

\item{progress}{An boolean that specifies whether a progress bar should display for each run of the optimization. This works only when multicore = F. DEFAULT = TRUE.}

\item{multicore}{An boolean that specifies whether the program should be run using multiple processing cores. DEFAULT = FALSE.}

\item{multicorePackages}{An vector of strings specifying the package names that are used in "fn." This needs to be used when multicore = TRUE. DEFAULT = NULL.}
}
\value{
A list containing: (1) currentBest (the parameter values for the best fit) and (2) best10 (the best fit parameter values for the top optParamListN runs)
}
\description{
This is a "smart" grid search optimization function. For each parameter, it randomly selects valid values within a set of bounds.  It then runs the model to be optimized and saves the fit statistic.  The smart grid search optimization repeats this numLoops number of times (Default = 200).  The smart grid search then reduces the bounds of each parameter based on the spread of the parameters in the top N fits (parameter = optParamListN, Default = 10).  This procedure repeats until the fit of the model fails to improve.  This optimization routine accurately identified parameters on simulated data when extensively tested.
}
\examples{
smartGridSearch (assessRRWfit, parsUpper = list(a = (a1 = 1, a2 = 1)), parsLower = list(a = (a1 = 0, a2 = 0)), parsMinInt = list(a = (a1 = 0.01, a2 = 0.01)), otherParamList = list( data = data.frame ( ... ), ...), numLoops = 100, numIntervals = 100, optBoundLoops = 10, multicore = TRUE, multicorePackages = c('RRW'))
}
\keyword{Grid}
\keyword{Search}
\keyword{Smart}
\keyword{optimize}
